Move VehicleWaterPlacementArgs to VehiclePlaceWater.h
