+++ src/OpenLoco/src/GameCommands/GameCommands.cpp
@@ -79,6 +79,7 @@
 #include "Vehicles/VehiclePickupWater.h"
 #include "Vehicles/VehiclePlace.h"
 #include "Vehicles/VehiclePlaceAir.h"
+#include "Vehicles/VehiclePlaceWater.h"
 #include "Vehicles/VehicleRearrange.h"
 #include "Vehicles/VehicleRefit.h"
 #include "Vehicles/VehicleReverse.h"
