+++ src/OpenLoco/src/GameCommands/GameCommands.h
@@ -182,34 +182,6 @@ namespace OpenLoco::GameCommands
         }
     };
 
-    struct VehicleWaterPlacementArgs
-    {
-        static constexpr auto command = GameCommand::vehiclePlaceWater;
-
-        VehicleWaterPlacementArgs() = default;
-        explicit VehicleWaterPlacementArgs(const registers& regs)
-            : pos(regs.ax, regs.cx, regs.dx)
-            , head(EntityId(regs.di))
-            , convertGhost((regs.ebx >> 16) == 0xFFFF)
-        {
-        }
-
-        World::Pos3 pos;
-        EntityId head;
-        bool convertGhost = false;
-
-        explicit operator registers() const
-        {
-            registers regs;
-            regs.ax = pos.x;
-            regs.cx = pos.y;
-            regs.dx = pos.z;
-            regs.di = enumValue(head);
-            regs.ebx = convertGhost ? 0xFFFF0000 : 0;
-            return regs;
-        }
-    };
-
     // Load multiplayer map
     inline void do_67(const char* filename)
     {
