+++ src/OpenLoco/src/GameCommands/Vehicles/VehiclePlaceWater.h
@@ -1,0 +1,34 @@
+#pragma once
+
+#include "GameCommands/GameCommands.h"
+
+namespace OpenLoco::GameCommands
+{
+    struct VehicleWaterPlacementArgs
+    {
+        static constexpr auto command = GameCommand::vehiclePlaceWater;
+
+        VehicleWaterPlacementArgs() = default;
+        explicit VehicleWaterPlacementArgs(const registers& regs)
+            : pos(regs.ax, regs.cx, regs.dx)
+            , head(EntityId(regs.di))
+            , convertGhost((regs.ebx >> 16) == 0xFFFF)
+        {
+        }
+
+        World::Pos3 pos;
+        EntityId head;
+        bool convertGhost = false;
+
+        explicit operator registers() const
+        {
+            registers regs;
+            regs.ax = pos.x;
+            regs.cx = pos.y;
+            regs.dx = pos.z;
+            regs.di = enumValue(head);
+            regs.ebx = convertGhost ? 0xFFFF0000 : 0;
+            return regs;
+        }
+    };
+}
